pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yurtle-kanban"
version = "0.1.0"
description = "File-based kanban using Yurtle (Turtle RDF in Markdown). Git is your database."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Hank Hancock", email = "hank@example.com" }
]
keywords = [
    "kanban",
    "yurtle",
    "rdf",
    "turtle",
    "workflow",
    "project-management",
    "cli",
    "mcp"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "rdflib>=7.0.0",
    "yurtle-rdflib>=0.1.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
mcp = [
    "mcp>=0.1.0",
]

[project.scripts]
yurtle-kanban = "yurtle_kanban.cli:main"

[project.urls]
Homepage = "https://github.com/hankh95/yurtle-kanban"
Documentation = "https://github.com/hankh95/yurtle-kanban#readme"
Repository = "https://github.com/hankh95/yurtle-kanban"
Issues = "https://github.com/hankh95/yurtle-kanban/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/yurtle_kanban"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
